Make blog posts on index be equal height
We're going to use flexbox here because it's awesome and we're making an
assumption that people reading this blog are developers that are using
modern web browsers.

We're able to remove the `.article-list-item-heading` div as it was only
being used to create the border effect which we can now just put on the
article itself.

Flexbox!