Show banner at bottom of page when not on homepage
The homepage has a nice dark content area, so a bold orange banner
looks good there and we want the conference alert to be an
important announcement.

Other pages have small, light headers. So although we could simply
change the style of the banner on these other pages, it's also
a little annoying to have it at the top of the page forever.