Fix heart animation bug
When the user moved the mouse off the element and then back again before
the animation finished, the heartbeat would stop (and wouldn't start
again). This was because our event listeners to stop the animation were
still present. To fix this, the event listeners to stop the animation
have are now cleared on mouse over.