Fix off center mobile view
This bug was introduced when additional padding was added around the
menu toggle button (mobile). The extra padding moved the button 12px to
the left, which made the button look too far away from the edge of the
screen. The previous solution adjusted the padding around the
navbar to accomodate for this, but it also inadvertently changed the
padding for the entire layout. This new fix reverts that change and
applys the 12px margin change directly to the menu element itself.