Fix image jumping
When switching tabs in the two tab panes on the homepage, the image
would move around. This was caused by the text in one column
growing/shrinking, which then changed the size of the div, resulting in
the image needing to be centered again. To fix this, a fixed height
value has been set.