Add template for a GitHub commit element
This template represents the visual aspect of a GitHub commit. It is
intended to look exectly like a commit would on GitHub. As it is a
template, the browser won't render it; we will use some JS to create
copies of this element and populate them with data from the GitHub API.