Render live commits from GitHub API
By getting commit data directly from GitHub, we can display a live
version of the latest GitHub commits. This is superior to an image as it
will always be up to date. It will also scale much better with higher
resolution displays as we're using elements that can be scaled by the
browser (as opposed to a static image).