Fix GitHub commit title parsing error
When parsing the response from GitHub, I made the assumption that a
commit with the author of 'webflow' would always be a merge commit; this
is not the case.

To fix this, I've added an extra check to make sure that the commit
message is actually a merge commit by checking that the commit message
starts with 'Merge'.