Retrieve GitHub contributors via GitHub APIs
It uses v4 (GraphQL API) to get exactly what we need with one request.

It converts the result in something easily manageable and exandable if
we have other needs in the future.

Contributor class implements methods that allow comparing and pushing
object to a SortedSet. This allows to have sorting and uniqueness of
element in the collection for free. I just changed how add works on the
Contributors class in order to increment the contributions count on that
user if it is already present.