Animate increment of stats count in homepage
To trigger the animation we need to subscribe to an event
'aos:in' fired by aos when objects animation start.

Unfortunately, this feature is not available in stable aos
versions, and an update of this package to a beta version is
required.

The alternative is playing with mutation observers with a ton
of new code or switch to other JS libraries.